adding statement generator baseline
