+++ generative_social_choice/statements/statement_generation.py
@@ -5,8 +5,39 @@ from typing import override, Tuple, List
 
 import pandas as pd
 
-from generative_social_choice.queries.query_interface import Generator, Agent
+from generative_social_choice.queries.query_interface import Generator,Agent
 from generative_social_choice.utils.gpt_wrapper import LLMLog
+from generative_social_choice.queries.query_chatbot_personalization import ChatbotPersonalizationGenerator
+
+
+class SimplePersonalizationAgent(Agent):
+    """Simple agent representation which doesn't require connecting to any LLM
+    but can't be used to get approvals.
+    
+    We use this class in statement generation as computing new approvals
+    is unnecessary."""
+
+    def __init__(
+        self,
+        *,
+        id: str,
+        survey_responses: pd.DataFrame,
+        summary: str,
+    ):
+        self.id = id
+        self.survey_responses = survey_responses
+        self.summary = summary
+
+    def get_id(self):
+        return self.id
+
+    def get_description(self):
+        return self.summary
+
+    def get_approval(
+        self, statement: str, use_logprobs: bool = True
+    ) -> tuple[float, list[LLMLog]]:
+        raise NotImplementedError()
 
 
 class NamedGenerator(Generator):
@@ -40,3 +71,20 @@ class DummyGenerator(NamedGenerator):
             new_statement = ''.join(random.choices(string.ascii_letters + " ", k=self.statement_length))
             statements.append(new_statement)
         return statements, []
+    
+
+class NamedChatbotPersonalizationGenerator(ChatbotPersonalizationGenerator, NamedGenerator):
+    """Use an LLM to generate a new statement for the given agents.
+    
+    This class is very similar to queries.query_chatbot_personalization.ChatbotPersonalizationGenerator,
+    but has the additional name property for logging"""
+    def __init__(
+        self,
+        *,
+        seed: int | None = None,
+        gpt_temperature=0.0,
+        model: str = "gpt-4o-mini-2024-07-18",
+    ):
+        init_args = {"seed": seed, "gpt_temperature": gpt_temperature, "model": model}
+        super().__init__(**init_args)
+        self._init_args = init_args
